MAKE SURE TO POST EVERY PROPERTY OF A MODEL WHEN EDITING COLUMNS
For any method that deals with editing, we need to POST ALL OF THE COLUMNS of the Model using  <input type="hidden" asp-for="[MODELPROPERTYNAME]" /> in your razor view file or else any column with existing data that isn't edited will turn out NULL!! Oh and you also have to change your Binding attributes to all columns of the table (since you're POSTing every property of the Model).

I guess it somehow does this because when a user POSTs data to the controller, it expects to see every property to be changed, so those that aren't sent through are automatically set to null, so I came up with fixing this by writing the line above.

It basically reinputs the existing data for that column of that row but the type being hidden makes it that the user doesn't need to fill those up again , as it's already done out of sight.