Edit UserShiftStatuses ALMOST works (read description)
So basically the gist of why we couldn't Update this table like how we normally do with our other tables is because this relationship is a Many-To-Many WITH History; That's why there's dupes of the UserId and ShiftStatusId, so EF doesn't know what row to update when there's dupes.

The way I then did this is to then literally query the EXACT row (same user id, shift id, and most importantly TIME as that's the 'history' part) and basically do a delete and readd; Delete the original row then add in the new row with the user's updated values.

There is a stupid problem with this that I've explained in the comments of the Edit function when doing the query though: The entire thing breaks if our Time column has seconds with a value > 0 (so I can edit 11:45:00 PM but not 11:45:12 PM simply because the seconds is not 00), and I have no clue as to why. So we can either debug this, or just straight up cheese it and adjust all the times to set the seconds as 00.

Sorry for the messing coding I've literally was so confused as to why this was happening like it literally makes no sense :))))